backendv2/requirements.txt

torch==2.7.0+cpu

fastapi==0.104.1
uvicorn[standard]==0.24.0
httpx==0.25.1
anyio<4.0.0              
marker-pdf==1.6.2
redis==6.0.0
supabase==2.15.1
click==8.1.8
python-multipart==0.0.20